refactor(share): decompose ShareButton into focused sub-modules
Extract platform detection into useShareCapabilities, lightbox state
into useLightbox, and render subtrees into LightboxDialog /
CustomizePopover. ShareButton.tsx becomes a re-export barrel; visible
useState in the main component drops from 17 to 6.
